--- a/MapAnalyzer/__init__.py
+++ b/MapAnalyzer/__init__.py
@@ -1,4 +1,3 @@
-
 # flake8: noqa
 from .MapData import MapData
 from .Polygon import Polygon
@@ -7,6 +6,6 @@ from .constructs import ChokeArea, MDRamp, VisionBlockerArea
 from pkg_resources import get_distribution, DistributionNotFound
 
 try:
-    __version__ = get_distribution('sc2mapanalyzer')
+    __version__ = get_distribution("sc2mapanalyzer")
 except DistributionNotFound:
-    __version__ = 'dev'
+    __version__ = "dev"