No rush to merge this, feedback is appreciated. These changes are currently live in Dysnomia on aiarena ladder

`MapData._compile_map` 
|Map |Previous(sec)| Current(sec)|
--- | --- | ---|
|2000AtmospheresAIE|1.7932|0.6596|
|BerlingradAIE|0.6749|0.2302|
|BlackburnAIE|1.3076|0.4268|
|CuriousMindsAIE|0.9635|0.2524|
|GlitteringAshesAIE|1.4703|0.4486|
|HardwireAIE|1.2705|0.3659|

Main improvement was found while calculating `Polygon` areas, which was previously calling `MapData.where_all` for every point on the perimeter and then checking if that point was inside any other choke or region etc
There is now an `extended_array_matrix` which contains the arrays for every `Polygon` created, using this we can check if a `Polygon''s perimeter points exist in other polygons in one operation.

Removed other blocks of code that looked like it was repeating this process of looking for subareas in earlier steps of the compile procedure. Please check this over, not sure if something was overlooked

Found another bottleneck while `Regions` were calculated, this was since it iterated all points and created a `Point2` for each one before working out the area. This is replaced with a quick call to `np.sum()` all the ones in the array