@raspersc2 2022-04-09
Reverted the last commit which I was unsure about anyhow, tests pass again locally at least. Workflow needs approval again

Will update timings in the original message later